io: move fObjMap-related methods
They are identical for TBufferFile and TBufferText
